[ Refactor ] : 전시 기록 작성 유스케이스 합치기
