+++ feature/home/src/main/java/com/yapp/gallery/home/screen/record/ExhibitRecordState.kt
@@ -14,4 +14,6 @@ sealed class ExhibitRecordState{
     // 기본 상태
     // 해당 상태에서 전시 기록할 시 서버로 요청 보내야 함
     object Normal : ExhibitRecordState()
+    // 이미 생성된 적 있는 상태
+    object Crated : ExhibitRecordState()
 }
